# web/openapi.json
{
  "components": {
    "schemas": {
...
      "EvaluationResult": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_s": {
            "title": "Duration S",
            "type": "number"
          },
          "gate_failures": {
            "items": {
              "type": "string"
            },
            "title": "Gate Failures",
            "type": "array"
          },
          "gate_passed": {
            "title": "Gate Passed",
            "type": "boolean"
          },
          "is_mock": {
            "title": "Is Mock",
            "type": "boolean"
          },
          "metrics": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Metrics",
            "type": "object"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "config",
          "metrics",
          "gate_passed",
          "duration_s",
          "is_mock"
        ],
        "title": "EvaluationResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReviewCorrectionRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "corrections": {
            "additionalProperties": true,
            "description": "Field name -> corrected value; may include 'document_type'. For answer reviews use {'answer': '<text>'}.",
            "title": "Corrections",
            "type": "object"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Placeholder reviewer identity; ignored when API-key auth is enabled.",
            "title": "Reviewer Id"
          }
        },
        "required": [
          "corrections"
        ],
        "title": "ReviewCorrectionRequest",
        "type": "object"
      },
      "ReviewDecisionRequest": {
        "additionalProperties": false,
        "properties": {
          "comment": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Placeholder reviewer identity; ignored when API-key auth is enabled.",
            "title": "Reviewer Id"
          }
        },
        "title": "ReviewDecisionRequest",
        "type": "object"
      },
      "ReviewListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "ReviewListResponse",
        "type": "object"
      },
      "ReviewReason": {
        "enum": [
          "low_classification_confidence",
          "unknown_document_type",
          "low_extraction_confidence",
          "schema_validation_failed",
          "validation_rule_failed",
          "conflicting_values",
          "missing_required_fields",
          "unverified_evidence",
          "insufficient_evidence",
          "weak_grounding",
          "low_answer_confidence",
          "guardrail_triggered",
          "retries_exhausted",
          "ocr_unavailable",
          "ocr_used"
        ],
        "title": "ReviewReason",
        "type": "string"
      },
      "ReviewResponse": {
        "properties": {
          "corrected_output": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Output"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "details": {
            "items": {
              "type": "string"
            },
            "title": "Details",
            "type": "array"
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "original_output": {
            "additionalProperties": true,
            "title": "Original Output",
            "type": "object"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "reasons": {
            "items": {
              "$ref": "#/components/schemas/ReviewReason"
            },
            "title": "Reasons",
            "type": "array"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "review_id": {
            "title": "Review Id",
            "type": "string"
          },
          "reviewer_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Comment"
          },
          "reviewer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Id"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "$ref": "#/components/schemas/ReviewTargetType"
          }
        },
        "required": [
          "review_id",
          "document_id",
          "target_type",
          "target_id",
          "status",
          "reasons",
          "details",
          "original_output",
          "corrected_output",
          "reviewer_id",
          "reviewer_comment",
          "model_version",
          "prompt_version",
          "created_at",
          "resolved_at"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
      "ReviewStatus": {
        "enum": [
          "pending",
          "approved",
          "rejected",
          "corrected",
          "superseded"
        ],
        "title": "ReviewStatus",
        "type": "string"
      },
      "ReviewTargetType": {
        "enum": [
          "document_processing",
          "answer"
        ],
        "title": "ReviewTargetType",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Ingestion, classification, extraction, validation, grounded Q&A with citations, human review, audit and evaluation for financial documents.",
    "title": "Financial Document Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/evaluations": {
      "get": {
        "operationId": "list_evaluations_evaluations_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationResult"
                  },
                  "title": "Response List Evaluations Evaluations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Evaluations",
        "tags": [
          "evaluations"
        ]
      }
    },
    "/evaluations/run": {
      "post": {
        "description": "Run the labelled evaluation suite in an isolated sandbox and store the result.\n\nSynchronous by design for the small synthetic dataset; with a real provider this makes\npaid model calls (roughly 3 per document plus 1 per question).",
        "operationId": "run_evaluation_evaluations_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Evaluation",
        "tags": [
          "evaluations"
        ]
      }
    },
...
    "/reviews": {
      "get": {
        "operationId": "list_reviews_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReviewStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "document_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Document Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reviews",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{review_id}": {
      "get": {
        "operationId": "get_review_reviews__review_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{review_id}/approve": {
      "post": {
        "operationId": "approve_review_reviews__review_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{review_id}/correct": {
      "post": {
        "operationId": "correct_review_reviews__review_id__correct_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCorrectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Correct Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/reviews/{review_id}/reject": {
      "post": {
        "operationId": "reject_review_reviews__review_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          },
          {
            "in": "cookie",
            "name": "docintel_demo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Docintel Demo"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reject Review",
        "tags": [
          "reviews"
        ]
      }
    }
  }
}